Add comments controller
